feat: 버튼에 join api 연결
